Add a MouseArea
Until touch is fixed up, touching on a touch-enabled monitor will only be
correct as long as the screen is the first in the virtual desktop layout.

Otherwise the mouse area on the first screen will be the one getting the
event. This can now be tested and visualized with the app.